Remove featuretools install command (#475)
* remove install

* remove more unused code from tests

* add test and more clean up

* move functions from install_primitives.py to utils.py

* remove additional files related to primitive installation tests

* remove s3fs dependency
