+++ featuretools/__init__.py
@@ -5,7 +5,7 @@ from . import variable_types
 from .entityset.api import *
 from . import primitives
 from .synthesis.api import *
-from .primitives import list_primitives, install_primitives
+from .primitives import list_primitives
 from .computational_backends.api import *
 from . import tests
 from .utils.pickle_utils import *
