+++ featuretools/__main__.py
@@ -2,8 +2,7 @@ import click
 import pandas as pd
 
 import featuretools
-from featuretools.primitives import install_primitives
-from featuretools.primitives.install import get_featuretools_root
+from featuretools.primitives.utils import get_featuretools_root
 
 
 @click.group()
@@ -18,19 +17,11 @@ def info():
 
 
 @click.command()
-@click.option('--prompt/--no-prompt', default=True, help='Confirm primitives before installing')
-@click.argument('directory')
-def install(prompt, directory):
-    install_primitives(directory, prompt)
-
-
-@click.command()
 def list_primitives():
     with pd.option_context('display.max_rows', None, 'display.max_columns', None, 'display.max_colwidth', -1, 'display.width', 1000):
         print(featuretools.list_primitives())
 
 
-cli.add_command(install)
 cli.add_command(list_primitives)
 cli.add_command(info)
