+++ featuretools/primitives/api.py
@@ -1,6 +1,4 @@
 # flake8: noqa
-from .install import install_primitives
-from .installed import *
 from .standard.aggregation_primitives import *
 from .standard.binary_transform import *
 from .standard.cum_transform_feature import *
