--- featuretools/tests/primitive_tests/bad_primitive_files/info.json
@@ -1,1 +1,0 @@
-{"primitives": ["CustomMax", "CustomSum"]}
