--- featuretools/tests/primitive_tests/primitives_to_install/data/_pytest_test.csv
@@ -1,3 +1,0 @@
-test
-file
-text
