--- featuretools/tests/primitive_tests/primitives_to_install/data/pytest_folder/hello.txt
@@ -1,1 +1,0 @@
-hello world
