--- featuretools/tests/primitive_tests/primitives_to_install/featuretools_pip_tester/setup.py
@@ -1,7 +1,0 @@
-from setuptools import setup
-
-setup(name='featuretools-pip-tester',
-      version='0.1',
-      description='Module to test if the featuretools install commmand will '
-                  'install pip modules',
-      packages=['featuretools_pip_tester'])
