--- featuretools/tests/primitive_tests/primitives_to_install/info.json
@@ -1,1 +1,0 @@
-{"primitives": ["CustomMax", "CustomMean", "CustomSum"]}
