+++ requirements.txt
@@ -1,5 +1,3 @@
-boto3>=1.9.51
-botocore>=1.12.51
 numpy>=1.13.3
 pandas>=0.23.0
 tqdm>=4.19.2
@@ -11,5 +9,4 @@ dask>=1.1.0
 distributed>=1.24.2
 psutil>=5.4.8
 click>=7.0.0
-smart_open>=1.7.1
 scikit-learn>=0.20.0
